fix key encoding issues, request mgmt tests
