+++ packages/contracts/test/helpers/testSetup.ts
@@ -2,7 +2,34 @@ import { ethers } from 'hardhat';
 import hre from 'hardhat';
 import ignitionModules from '../../ignition/modules';
 import type { AccountFactory, AccountRegistry, TestContract } from '../../typechain-types';
-import { generateTestKeypair } from './WebAuthnSignatures';
+import { createKeypair } from '../utils';
+import crypto from 'node:crypto'
+
+export type HexifiedPublicKey = {
+  x: string
+  y: string
+}
+
+export type KeyPair = {
+  publicKey: HexifiedPublicKey,
+  keyPair: crypto.KeyPairKeyObjectResult
+}
+
+/**
+ * Generate a testing keypair for WebAuthn signatures
+ * @returns The keypair with formatted public key for Solidity
+ */
+export function generateTestKeypair(): KeyPair {
+  const { x, y, keyPair } = createKeypair();
+
+  // Format the public key for Solidity contracts
+  const publicKey = {
+    x: ethers.hexlify(x),
+    y: ethers.hexlify(y),
+  };
+
+  return { publicKey, keyPair };
+}
 
 /**
  * Deploy contracts fixture
